requirements.txt
alembic==1.9.3
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.91.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.4
PyJWT==1.7.1
sniffio==1.3.0
SQLAlchemy==2.0.3
SQLAlchemy-Utils==0.40.0
starlette==0.24.0
typing_extensions==4.4.0
uvicorn==0.20.0
Werkzeug==2.2.3
kafka-python
requests